pyproject.toml
[tool.poetry]
name = "dataanalysisvisualizationfiles"
version = "0.1.0"
description = ""
authors = ["Yarden4998 <yardenk@campus.technion.ac.il>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
numpy = "1.26.4"
pandas = "^2.2.2"
matplotlib = "^3.9.1.post1"
seaborn = "^0.13.2"
faiss-cpu = "^1.8.0.post1"
scikit-learn = "^1.5.1"
scipy = "^1.14.0"
ipykernel = "^6.29.5"
nltk = "^3.8.2"
wordcloud = "^1.9.3"
torch = "2.2.2"
tensorflow-io-gcs-filesystem = "0.31.0"
tensorflow = "2.15.0"
torchvision = "0.17.2"
sentence-transformers = "^3.0.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
